tracing/uprobe: Add missing filter for uretprobes
Uprobes are designed to be filterable using the PID, which is the second
parameter in the perf_event_open syscall. Currently, uprobe works well
with the filtering, but uretprobe is not affected by it. This often leads
to users being disturbed by events from uninterested processes while using
uretprobe.

We found that the filter function was not invoked when uretprobe was
initially implemented, and this has been existing for ten years. We have
tested the patch under our workload, binding eBPF programs to uretprobe
tracepoints, and confirmed that it resolved our problem.

Fixes: c1ae5c75e103 ("uprobes/tracing: Introduce is_ret_probe() and uretprobe_dispatcher()")
Co-developed-by: Francis Laniel <flaniel@linux.microsoft.com>
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
Cc: Alban Crequy <albancrequy@linux.microsoft.com>
Signed-off-by: Tianyi Liu <i.pear@outlook.com>